Dissertation python3 code V0.1
This code can be opened in Google Colab as a notebook.
The code consists of links to the datasets and their authors.
In addition, it also consists of fully working Feed Forward Neural Network and Convolutional Neural Network which were applied to Fruits-360 dataset only. Further work to be done.