Uploading the second version of the code
This version, does not add any new features to the code itself. It just adds a bit more comments, removes unnecessary lines which were not needed.
The main purpose of this upload is to see the accuracy of FNN and CNN on the provided dataset, which is a different one from the DissertationCode_V0.1.ipynb. The dataset used is a "Waste Classification" dataset, which is a combination of two datasets.